ai-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.39.0
openai==1.58.1
google-generativeai==0.8.3
redis==5.2.1
PyJWT==2.10.1
httpx==0.28.1
pydantic==2.10.3
pydantic-settings==2.6.1
sse-starlette==2.1.3
python-multipart==0.0.12
